Migration from LiveData to StateFlow

+++ app/src/main/java/com/android/tvmaze/favorite/FavoriteShowsViewModel.kt
@@ -1,21 +1,30 @@
 package com.android.tvmaze.favorite
 
 import androidx.hilt.lifecycle.ViewModelInject
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.android.tvmaze.db.favouriteshow.FavoriteShow
+import com.android.tvmaze.di.CoroutinesDispatcherProvider
 import kotlinx.coroutines.CoroutineExceptionHandler
-import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import timber.log.Timber
-import java.util.ArrayList
+import java.util.*
 
 class FavoriteShowsViewModel @ViewModelInject
-constructor(private val favoriteShowsRepository: FavoriteShowsRepository) : ViewModel() {
-    private val favoriteShowsLiveData: MutableLiveData<List<FavoriteShow>> = MutableLiveData()
+constructor(
+    private val favoriteShowsRepository: FavoriteShowsRepository,
+    // Inject coroutineDispatcher to facilitate Unit Testing
+    private val coroutinesDispatcherProvider: CoroutinesDispatcherProvider
+) : ViewModel() {
+    private val _favoriteShowsStateFlow =
+        MutableStateFlow<FavoriteShowState>(FavoriteShowState.Loading)
+
+    // Represents _favoriteShowsStateFlow mutable state flow as a read-only state flow.
+    val favoriteShowsStateFlow = _favoriteShowsStateFlow.asStateFlow()
+
     private lateinit var removedFromFavoriteShows: List<FavoriteShow>
 
     fun loadFavoriteShows() {
@@ -24,33 +33,35 @@ constructor(private val favoriteShowsRepository: FavoriteShowsRepository) : View
         }
 
         viewModelScope.launch(coroutineExceptionHandler) {
-            val favoriteShows = withContext(Dispatchers.IO + coroutineExceptionHandler) {
-                favoriteShowsRepository.allFavoriteShows()
-            }
-            withContext(Dispatchers.Main + coroutineExceptionHandler) {
-                onFavoritesFetched(favoriteShows)
+            val favoriteShows =
+                withContext(coroutinesDispatcherProvider.io) {
+                    favoriteShowsRepository.allFavoriteShows()
+                }
+            withContext(coroutinesDispatcherProvider.main) {
+                removedFromFavoriteShows = ArrayList(favoriteShows.size)
+                if (favoriteShows.isNotEmpty()) {
+                    _favoriteShowsStateFlow.emit(FavoriteShowState.AllFavorites(favoriteShows))
+                } else {
+                    _favoriteShowsStateFlow.emit(FavoriteShowState.Empty)
+                }
             }
         }
     }
 
     private fun onError(throwable: Throwable) {
+        _favoriteShowsStateFlow.value = FavoriteShowState.Error(throwable.localizedMessage)
         Timber.d(throwable)
     }
 
-    private fun onFavoritesFetched(favoriteShows: List<FavoriteShow>) {
-        removedFromFavoriteShows = ArrayList(favoriteShows.size)
-        favoriteShowsLiveData.value = favoriteShows
-    }
-
-    fun getFavoriteShowsLiveData(): LiveData<List<FavoriteShow>> {
-        return favoriteShowsLiveData
-    }
-
-    fun addToFavorite(show: FavoriteShow) {
-        favoriteShowsRepository.insertIntoFavorites(show)
-    }
-
-    fun removeFromFavorite(show: FavoriteShow) {
-        favoriteShowsRepository.removeFromFavorites(show)
+    fun onFavoriteClick(show: FavoriteShow) {
+        viewModelScope.launch(coroutinesDispatcherProvider.io) {
+            if (!show.isFavorite) {
+                favoriteShowsRepository.insertIntoFavorites(show)
+                _favoriteShowsStateFlow.emit(FavoriteShowState.AddedToFavorites(show))
+            } else {
+                favoriteShowsRepository.removeFromFavorites(show)
+                _favoriteShowsStateFlow.emit(FavoriteShowState.RemovedFromFavorites(show))
+            }
+        }
     }
 }
